test: Testing the Get Venues Nearby use case.

+++ Domain/src/main/java/com/baruckis/domain/model/Venue.kt
@@ -16,5 +16,5 @@
 
 package com.baruckis.domain.model
 
-class Venue(val name: String, val lattitude: Float, val longitude: Float) {
+class Venue(val name: String, val lattitude: Double, val longitude: Double) {
 }
